feat(forecaster): close FB-FR-P0 queue — MLP distill, torch float32, empty OPEN
- Add train_distilled_mlp_forecaster + nm-train-forecaster-distill CLI (teacher: NumPy ForecasterModel)
- describe_walk_forward_folds helper; tests for conformal sanity, torch→DecisionPipeline
- torch_infer: float32 inputs to match MLP weights; double numpy outputs preserved
- QUEUE.MD: archive FB-FR-P0, FB-SPEC-07, FB-MS; add FB-FR-P0-CLOSE/WAIVE notes
- MULTI_ASSET_PORTFOLIO.MD: exposure + independent-per-symbol + operator warnings
- README/RUNBOOKS: document distill entrypoint and serving precedence

+++ forecaster_model/inference/torch_infer.py
@@ -64,8 +64,8 @@ def forward_torch_quantiles(
 
     model.eval()
     with torch.no_grad():
-        xo = torch.from_numpy(np.asarray(x_obs, dtype=np.float64)).unsqueeze(0).to(device)
-        xk = torch.from_numpy(np.asarray(x_known, dtype=np.float64)).unsqueeze(0).to(device)
-        rr = torch.from_numpy(np.asarray(r_cur, dtype=np.float64).reshape(1, -1)).to(device)
+        xo = torch.from_numpy(np.asarray(x_obs, dtype=np.float32)).float().unsqueeze(0).to(device)
+        xk = torch.from_numpy(np.asarray(x_known, dtype=np.float32)).float().unsqueeze(0).to(device)
+        rr = torch.from_numpy(np.asarray(r_cur, dtype=np.float32).reshape(1, -1)).float().to(device)
         out = model(xo, xk, rr)
-        return out.squeeze(0).detach().cpu().numpy().astype(np.float64)
+        return out.squeeze(0).detach().cpu().numpy().astype(np.float64, copy=False)
